Rename DefinitionTypeHelpers to ResourceTypes
