+++ packages/meteor-type-validation/src/server/index.ts
@@ -1,4 +1,4 @@
 export * from './guards';
 export * from './ShorthandFunctions';
 export * from './MeteorTypeValidation';
-export type * as DefinitionTypeHelpers from './DefinitionTypeHelpers';
+export type * as DefinitionTypeHelpers from './ResourceTypes';
